+++ addons/test_mail_full/__manifest__.py
@@ -11,8 +11,6 @@ and mail-related sub modules. Those tests are present in a separate module as it
 contains models used only to perform tests independently to functional aspects of
 real applications. """,
     'depends': [
-        'test_mail',
-        'test_mass_mailing',
         'mail',
         'mail_bot',
         'portal',
@@ -22,6 +20,9 @@ real applications. """,
         'mass_mailing_sms',  # adds portal
         'phone_validation',
         'sms',
+        'test_mail',
+        'test_mail_sms',
+        'test_mass_mailing',
     ],
     'data': [
         'data/mail_message_subtype_data.xml',
