+++ addons/test_mail_full/models/__init__.py
@@ -1,4 +1,3 @@
 # -*- coding: utf-8 -*-
 
 from . import test_mail_models_mail
-from . import test_mail_models_sms
