+++ addons/test_mail_full/tests/__init__.py
@@ -1,19 +1,9 @@
 # -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import common
 from . import test_odoobot
 from . import test_mail_performance
-from . import test_phone_blacklist
 from . import test_mass_mailing
-from . import test_mass_mailing_statistics
-from . import test_mass_sms
 from . import test_portal
 from . import test_rating
 from . import test_res_users
-from . import test_sms_composer
-from . import test_sms_management
-from . import test_sms_performance
-from . import test_sms_post
-from . import test_sms_server_actions
-from . import test_sms_sms
-from . import test_sms_template
