+++ addons/test_mail_full/tests/test_res_users.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.mail.tests.common import MailCommon, mail_new_test_user
+from odoo.addons.mail.tests.common import mail_new_test_user
+from odoo.addons.test_mail_full.tests.common import TestMailFullCommon
 
 
-class TestResUsers(MailCommon):
+class TestResUsers(TestMailFullCommon):
 
     @classmethod
     def setUpClass(cls):
