+++ addons/test_mail_sms/__manifest__.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+
+{
+    'name': 'SMS Tests',
+    'version': '1.0',
+    'category': 'Hidden',
+    'sequence': 9876,
+    'summary': 'SMS Tests: performances and tests specific to SMS',
+    'description': """This module contains tests related to SMS. Those are
+present in a separate module as it contains models used only to perform
+tests independently to functional aspects of other models. """,
+    'depends': [
+        'mail',
+        'sms',
+        'test_performance',
+    ],
+    'data': [
+        'security/ir.model.access.csv',
+    ],
+    'installable': True,
+    'license': 'LGPL-3',
+}
