+++ addons/test_mail_sms/tests/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from . import test_phone_blacklist
+from . import test_sms_composer
+from . import test_sms_management
+from . import test_sms_performance
+from . import test_sms_post
+from . import test_sms_server_actions
+from . import test_sms_sms
+from . import test_sms_template
