+++ addons/test_mail_sms/tests/test_phone_blacklist.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.test_mail_full.tests.common import TestMailFullCommon, TestMailFullRecipients
+from odoo.addons.test_mail_sms.tests.common import TestSMSCommon, TestSMSRecipients
 
 
-class TestPhoneBlacklist(TestMailFullCommon, TestMailFullRecipients):
-    """ TODO """
+class TestPhoneBlacklist(TestSMSCommon, TestSMSRecipients):
+    """ Test phone blacklist management """
 
     @classmethod
     def setUpClass(cls):
