+++ addons/test_mail_sms/tests/test_sms_server_actions.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.test_mail_full.tests.common import TestMailFullCommon, TestMailFullRecipients
+from odoo.addons.test_mail_sms.tests.common import TestSMSCommon, TestSMSRecipients
 
 
-class TestServerAction(TestMailFullCommon, TestMailFullRecipients):
+class TestServerAction(TestSMSCommon, TestSMSRecipients):
 
     @classmethod
     def setUpClass(cls):
