+++ addons/test_mass_mailing/__manifest__.py
@@ -9,7 +9,12 @@
     'description': """This module contains tests related to mass mailing. Those
 are present in a separate module to use specific test models defined in
 test_mail. """,
-    'depends': ['test_mail', 'mass_mailing'],
+    'depends': [
+        'mass_mailing',
+        'mass_mailing_sms',
+        'test_mail',
+        'test_mail_sms',
+    ],
     'data': [
         'security/ir.model.access.csv',
     ],
