+++ addons/test_mass_mailing/tests/test_mailing_statistics_sms.py
@@ -4,13 +4,13 @@
 from lxml import html
 
 from odoo.addons.mail.tests.common import mail_new_test_user
-from odoo.addons.test_mail_full.tests.test_mass_sms import TestMassSMSCommon
+from odoo.addons.test_mass_mailing.tests.common import TestMassSMSCommon
 from odoo.tests.common import users
 from odoo.tests import tagged
 from odoo.tools import mute_logger
 
 
-@tagged('digest', 'mass_mailing')
+@tagged('digest', 'mass_mailing', 'mass_mailing_sms')
 class TestMailingStatistics(TestMassSMSCommon):
 
     @classmethod
